+++ NEWS.md
@@ -4,6 +4,7 @@
   - `set_class()`: A pipeable version of `class(x) <-`
   - `deframe_with_class()`: First calls `tibble::deframe()` on a data.frame, then calls `set_class()` on each element of the resulting vector
 - Add `enumerate()` which returns the number of unique values found in a column of a data.frame
+- Add `paste_conjunction()` which collapses a vector with commas as well as a conjunction
 
 # glebrt 0.2.1